+++ changelog.md
@@ -2,6 +2,7 @@
 - Project is now buildable by poetry. `requirements.txt` and `setup.py` are removed.
 - Added `YOLOv8_ModelSpec`
 - Added pydantic2+ version support
+- Updated pathy's requirements.
 
 # 0.16.8
 - Fix `get_image_size` when filepath is bytes type